Create Stars component, bugfixes, stars functionality, small UI adjustments
