+++ client/src/assets/related-and-comparisons.sass
@@ -188,6 +188,7 @@
         flex-direction: column
         justify-content: space-between
         align-items: center
+        cursor: pointer
         padding-top: 10px
         transition: 'opacity 500ms ease-in, visibility 0ms ease-in 250ms'
         opacity: 1
