+++ client/src/components/RelatedAndComparisons/Card.jsx
@@ -4,11 +4,15 @@ import ProductInfo from './ProductInfo.jsx';
 import CurrentImage from './CurrentImage.jsx';
 
 const Cards = ( { product, setCurrentProduct, stars } ) => {
+  const [hoverToggle, setHoverToggle] = useState(false);
+
+  const hoverOver = () => setHoverToggle(true);
+  const hoverOut = () => setHoverToggle(false);
 
   return (
-    <div className='card' onClick={ () => setCurrentProduct(product.id)}>
+    <div className='card' onMouseEnter={hoverOver} onMouseLeave={hoverOut} onClick={ () => setCurrentProduct(product.id)}>
       <CurrentImage product={product} />
-      <ProductInfo setCurrentProduct={setCurrentProduct} product={product} stars={stars}/>
+      <ProductInfo setCurrentProduct={setCurrentProduct} product={product} stars={stars} hoverToggle={hoverToggle} />
     </div>
   );
